src/backend/musikk/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-debug-toolbar==5.1.0
django-jazzmin==3.0.1
iniconfig==2.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pytest==8.3.5
pytest-django==4.10.0
python-decouple==3.8
python-magic==0.4.28
sqlparse==0.5.3
typing_extensions==4.12.2
